Refactor manual math transforms
