+++ src/core/src/ast/index.ts
@@ -1,6 +1,7 @@
 export * from "./location-keys.js";
 export * from "./locations.js";
 export * from "./location-manipulation.js";
+export * from "./node-types.js";
 export * from "./node-helpers.js";
 export * from "./object-graph.js";
 export type {
