+++ src/core/src/ast/node-types.ts
@@ -1,0 +1,14 @@
+/**
+ * Shared AST node-type string constants used across parser transforms.
+ */
+export const ASSIGNMENT_EXPRESSION = "AssignmentExpression";
+export const BINARY_EXPRESSION = "BinaryExpression";
+export const CALL_EXPRESSION = "CallExpression";
+export const EXPRESSION_STATEMENT = "ExpressionStatement";
+export const IDENTIFIER = "Identifier";
+export const LITERAL = "Literal";
+export const MEMBER_DOT_EXPRESSION = "MemberDotExpression";
+export const MEMBER_INDEX_EXPRESSION = "MemberIndexExpression";
+export const PARENTHESIZED_EXPRESSION = "ParenthesizedExpression";
+export const UNARY_EXPRESSION = "UnaryExpression";
+export const VARIABLE_DECLARATION = "VariableDeclaration";
